refactor: resolve project dep func
